optimize LZ77 encoding

+++ src/io/buffered_writer.hpp
@@ -1,13 +1,15 @@
 #include "zemita/utils.hpp"
 
-class BufferedWriter{
-public:    
+class BufferedWriter
+{
+   public:
     explicit BufferedWriter(const std::string& out, const uint32_t bufferSize);
     void write(const char* toBeWrittenData, size_t dataSize);
     void flush();
-private:
+
+   private:
     std::ofstream out_;
     uint32_t bufferSize_;
     std::vector<char> buffer_;
-    uint16_t writePos_ = 0;
+    uint32_t writePos_ = 0;
 };
